VLab Issue #54494 - CAVE Freezes while loading RBDs
10/09/2018:
Made the modifications below to utilize existing JobPool
A. Added new 'processData()' method with logic that was previously in 'NcgridLoaderTask.run()'.

B. Moved logic from the 'NcgridResource.processNewRscDataList()' method to 'NcgridLoaderTask.run()'.

C. Took out the call to 'ncgridLoaderPool.cancel(ncgridLoaderTask)' from the 'processNewRscDataList()' method.

10/11/2018:
A. Updated the 'processData()' method to include the 'if(isFirst)'
statement and the call to the 'issueRefresh()' method.

Change-Id: I056a38793fd8452b3b1141908e8c23937211db0f